Extract out new module `TokenName`.
