+++ lib/primitive/lib/Cardano/Wallet/Primitive/Types/TokenBundle.hs
@@ -89,9 +89,11 @@ import Cardano.Wallet.Primitive.Types.TokenMap
     , Nested (..)
     , TokenMap
     )
-import Cardano.Wallet.Primitive.Types.TokenPolicy
+import Cardano.Wallet.Primitive.Types.TokenName
     ( TokenName
-    , TokenPolicyId
+    )
+import Cardano.Wallet.Primitive.Types.TokenPolicy
+    ( TokenPolicyId
     )
 import Cardano.Wallet.Primitive.Types.TokenQuantity
     ( TokenQuantity (..)
