+++ lib/primitive/lib/Cardano/Wallet/Primitive/Types/TokenPolicy.hs
@@ -13,12 +13,6 @@ module Cardano.Wallet.Primitive.Types.TokenPolicy
       -- * Token Policies
       TokenPolicyId (..)
 
-      -- * Token Names
-    , TokenName (..)
-    , mkTokenName
-    , nullTokenName
-    , tokenNameMaxLength
-
       -- * Token Fingerprints
     , TokenFingerprint (..)
     , mkTokenFingerprint
@@ -40,6 +34,9 @@ import Prelude
 
 import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..)
+    )
+import Cardano.Wallet.Primitive.Types.TokenName
+    ( TokenName (..)
     )
 import Codec.Binary.Bech32.TH
     ( humanReadablePart
@@ -65,11 +62,6 @@ import Data.Bifunctor
     )
 import Data.ByteArray
     ( convert
-    )
-import Data.ByteArray.Encoding
-    ( Base (Base16)
-    , convertFromBase
-    , convertToBase
     )
 import Data.ByteString
     ( ByteString
@@ -106,7 +98,6 @@ import Quiet
 import qualified Codec.Binary.Bech32 as Bech32
 import qualified Data.ByteString as BS
 import qualified Data.Text as T
-import qualified Data.Text.Encoding as T
 
 -- | Token policy identifiers, represented by the hash of the monetary policy
 -- script.
@@ -133,57 +124,6 @@ instance ToText TokenPolicyId where
 
 instance FromText TokenPolicyId where
     fromText = fmap UnsafeTokenPolicyId . fromText
-
--- | Token names, defined by the monetary policy script.
-newtype TokenName =
-    -- | Construct a 'TokenName' without any validation.
-    UnsafeTokenName { unTokenName :: ByteString }
-    deriving stock (Eq, Ord, Generic)
-    deriving (Read, Show) via (Quiet TokenName)
-    deriving anyclass Hashable
-
--- | Construct a 'TokenName', validating that the length does not exceed
---   'tokenNameMaxLength'.
---
-mkTokenName :: ByteString -> Either String TokenName
-mkTokenName bs
-    | BS.length bs <= tokenNameMaxLength = Right $ UnsafeTokenName bs
-    | otherwise = Left $ "TokenName length " ++ show (BS.length bs)
-        ++ " exceeds maximum of " ++ show tokenNameMaxLength
-
--- | The empty asset name.
---
--- Asset names may be empty, where a monetary policy script only mints a single
--- asset, or where one asset should be considered as the "default" token for the
--- policy.
---
-nullTokenName :: TokenName
-nullTokenName = UnsafeTokenName ""
-
--- | The maximum length of a valid token name.
---
-tokenNameMaxLength :: Int
-tokenNameMaxLength = 32
-
-instance NFData TokenName
-
-instance Buildable TokenName where
-    build = build . toText
-
-instance FromJSON TokenName where
-    parseJSON = parseJSON >=> either (fail . show) pure . fromText
-
-instance ToJSON TokenName where
-    toJSON = toJSON . toText
-
-instance ToText TokenName where
-    toText = T.decodeLatin1 . convertToBase Base16 . unTokenName
-
-instance FromText TokenName where
-    fromText = first TextDecodingError
-        . either (Left . ("TokenName is not hex-encoded: " ++)) mkTokenName
-        . convertFromBase Base16
-        . T.encodeUtf8
 
 newtype TokenFingerprint =
     UnsafeTokenFingerprint { unTokenFingerprint :: Text }
