+++ lib/primitive/test/spec/Cardano/Wallet/Primitive/Types/TxSpec.hs
@@ -27,9 +27,11 @@ import Cardano.Wallet.Primitive.Types.TokenMap.Gen
     ( genAssetId
     , shrinkAssetId
     )
-import Cardano.Wallet.Primitive.Types.TokenPolicy
+import Cardano.Wallet.Primitive.Types.TokenName
     ( TokenName (..)
-    , TokenPolicyId (..)
+    )
+import Cardano.Wallet.Primitive.Types.TokenPolicy
+    ( TokenPolicyId (..)
     )
 import Cardano.Wallet.Primitive.Types.Tx
     ( SealedTx (..)
