+++ lib/primitive/test/spec/Cardano/Wallet/Primitive/Types/UTxOSpec.hs
@@ -33,9 +33,11 @@ import Cardano.Wallet.Primitive.Types.TokenMap.Gen
     ( genAssetId
     , shrinkAssetId
     )
-import Cardano.Wallet.Primitive.Types.TokenPolicy
+import Cardano.Wallet.Primitive.Types.TokenName
     ( TokenName (..)
-    , TokenPolicyId (..)
+    )
+import Cardano.Wallet.Primitive.Types.TokenPolicy
+    ( TokenPolicyId (..)
     )
 import Cardano.Wallet.Primitive.Types.Tx.TxIn
     ( TxIn (..)
