+++ lib/wallet/api/http/Cardano/Wallet/Api/Types/MintBurn.hs
@@ -42,9 +42,11 @@ import Cardano.Wallet.Api.Types.Primitive
 import Cardano.Wallet.Primitive.Types.TokenMap
     ( toNestedList
     )
-import Cardano.Wallet.Primitive.Types.TokenPolicy
+import Cardano.Wallet.Primitive.Types.TokenName
     ( TokenName
-    , TokenPolicyId
+    )
+import Cardano.Wallet.Primitive.Types.TokenPolicy
+    ( TokenPolicyId
     , mkTokenFingerprint
     )
 import Cardano.Wallet.Primitive.Types.TokenQuantity
@@ -75,6 +77,7 @@ import Numeric.Natural
     )
 
 import qualified Cardano.Wallet.Primitive.Types.TokenMap as TokenMap
+import qualified Cardano.Wallet.Primitive.Types.TokenName as W
 import qualified Cardano.Wallet.Primitive.Types.TokenPolicy as W
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map.Strict as Map
