+++ lib/wallet/src/Cardano/Wallet/Primitive/Types/Tx/TxSeq.hs
@@ -125,9 +125,11 @@ import Cardano.Wallet.Primitive.Types.StateDeltaSeq
 import Cardano.Wallet.Primitive.Types.TokenMap
     ( AssetId (..)
     )
-import Cardano.Wallet.Primitive.Types.TokenPolicy
+import Cardano.Wallet.Primitive.Types.TokenName
     ( TokenName (..)
-    , TokenPolicyId (..)
+    )
+import Cardano.Wallet.Primitive.Types.TokenPolicy
+    ( TokenPolicyId (..)
     )
 import Cardano.Wallet.Primitive.Types.Tx
     ( Tx (..)
