+++ lib/wallet/src/Cardano/Wallet/TokenMetadata.hs
@@ -96,12 +96,14 @@ import Cardano.Wallet.Primitive.Types.Hash
 import Cardano.Wallet.Primitive.Types.TokenMap
     ( AssetId (..)
     )
+import Cardano.Wallet.Primitive.Types.TokenName
+    ( TokenName (..)
+    )
 import Cardano.Wallet.Primitive.Types.TokenPolicy
     ( AssetDecimals (..)
     , AssetLogo (..)
     , AssetMetadata (..)
     , AssetURL (..)
-    , TokenName (..)
     , TokenPolicyId (..)
     , validateMetadataDecimals
     , validateMetadataDescription
