+++ lib/wallet/test/unit/Cardano/Wallet/Api/Malformed.hs
@@ -101,9 +101,11 @@ import Cardano.Wallet.Primitive.Types
     ( WalletId
     , walletNameMaxLength
     )
-import Cardano.Wallet.Primitive.Types.TokenPolicy
+import Cardano.Wallet.Primitive.Types.TokenName
     ( TokenName
-    , TokenPolicyId
+    )
+import Cardano.Wallet.Primitive.Types.TokenPolicy
+    ( TokenPolicyId
     )
 import Cardano.Wallet.Primitive.Types.Tx
     ( SealedTx
