+++ lib/wallet/test/unit/Cardano/Wallet/Api/TypesSpec.hs
@@ -375,13 +375,15 @@ import Cardano.Wallet.Primitive.Types.TokenMap.Gen
     , genTokenMapSmallRange
     , shrinkTokenMap
     )
+import Cardano.Wallet.Primitive.Types.TokenName
+    ( TokenName (..)
+    )
 import Cardano.Wallet.Primitive.Types.TokenPolicy
     ( AssetDecimals (..)
     , AssetLogo (..)
     , AssetMetadata (..)
     , AssetURL (..)
     , TokenFingerprint
-    , TokenName (..)
     , TokenPolicyId (..)
     , mkTokenFingerprint
     )
